+++ modules/canvaskit/CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Changed
+### Fixed
+- `PathBuilder.setFillType` returns a reference to the JavaScript `PathBuilder` and does not make
+  a copy of the underlying SkPathBuilder.
+
+### Changed
 - `Canvas.drawImageCubic` and `Canvas.drawImageOptions` use
   `Fast_SrcRectConstraint` by default which allows the mipmap and filter options
   passed in to not be overwritten.